Refactor text spacing Expectation
